requirements.txt
django
django-admin-interface
django-debug-toolbar
whitenoise
openai
feedparser
pytz
BeautifulSoup4
tiktoken
gunicorn
fake_useragent
huey
redis
django-nested-admin
defusedxml
